move core functionalities to core.py
